django: Store `SetField` value in an SQL `text` column instead of `char(n)`
`SetField` stores values in a `char(n)` column. When a choice is added
or removed from the field, an SQL migration is generated because the
maximum length (to handle all choices) changes. This is annoying.
Using an unlimited column type avoids this issue. The performance of
the `text` and `char(n)` types are similar in PostgreSQL. I don't know
whether it's the same in other databases, but it's unlikely to be an
issue (please don't mark my words).